requirements.txt
google-adk>=1.2.0
mcp>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0